requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
slowapi>=0.1.9
rank-bm25>=0.2.2
pypdf>=3.17.0
docx2txt>=0.8
youtube-transcript-api>=0.6.0
webvtt-py>=0.4.6
srt>=3.5.3
authlib>=1.2.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0

# Database & Vector Store (pgvector)
psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0

# LLM Providers
openai>=1.0.0
anthropic>=0.7.0
httpx>=0.25.0

prometheus-fastapi-instrumentator>=7.1.0
stripe>=7.0.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-logging>=0.46b0

# Optional: Redis for caching (install if REDIS_ENABLED=true)
redis>=5.0.0

